System Support for Ubiquity-Distribute Tuple Space
The SysSU-DTS allows a transparent access to available context information,
through:
(i) local access, accesses the tuple space that is internal to the device,
(ii) infrastructured access to a tuple space located on a server via a network
infrastructured, or
(iii) Ad hoc access, interacting directly with tuple spaces located in nearby
devices via the formation of an Ad hoc network.

In addition, DTS-SysSU implements a "scoping mechanism" which permits the
formation of subsets of available contextual information. This scoping
mechanism to restrict access to contextual tuples only to members of the same
scope.